+++ knf/united/united_test.go
@@ -15,6 +15,8 @@ import (
 	"github.com/essentialkaos/ek/v12/knf"
 	"github.com/essentialkaos/ek/v12/options"
 
+	knfv "github.com/essentialkaos/ek/v12/knf/validators"
+
 	. "github.com/essentialkaos/check"
 )
 
@@ -217,4 +219,36 @@ func (s *UnitedSuite) TestWithEnv(c *C) {
 	c.Assert(GetTS("test:timestamp").Unix(), Equals, int64(1591014600))
 	c.Assert(GetTZ("test:timezone").String(), Equals, "Europe/Berlin")
 	c.Assert(GetL("test:list"), DeepEquals, []string{"Test1Env", "Test2Env"})
+}
+
+func (s *UnitedSuite) TestValidation(c *C) {
+	global = nil
+
+	errs := Validate([]*knf.Validator{
+		{"test:string", knfv.Empty, nil},
+	})
+
+	c.Assert(errs, HasLen, 1)
+
+	err := Combine(
+		Mapping{"test:string", "test-string", "TEST_STRING"},
+		Mapping{"test:integer", "test-integer", "TEST_INTEGER"},
+		Mapping{"test:float", "test-float", "TEST_FLOAT"},
+		Mapping{"test:boolean", "test-boolean", "TEST_BOOLEAN"},
+		Mapping{"test:file-mode", "test-file-mode", "TEST_FILE_MODE"},
+		Mapping{"test:duration", "test-duration", "TEST_DURATION"},
+		Mapping{"test:time-duration", "test-time-duration", "TEST_TIME_DURATION"},
+		Mapping{"test:timestamp", "test-timestamp", "TEST_TIMESTAMP"},
+		Mapping{"test:timezone", "test-timezone", "TEST_TIMEZONE"},
+		Mapping{"test:list", "test-list", "TEST_LIST"},
+	)
+
+	c.Assert(err, IsNil)
+
+	errs = Validate([]*knf.Validator{
+		{"test:string", knfv.Empty, nil},
+		{"test:integer", knfv.Greater, 100},
+	})
+
+	c.Assert(errs, HasLen, 1)
 }